renamed meta_db -> meta_version_control
